pyproject.toml
[project]
name = "ai-project01-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.12",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.1",
    "openai>=2.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
